moved asteroid file to components folder
