+++ react_game_project/game_core/src/components/Asteroid.jsx
